// src/types/radar.d.ts
export type AutoCompleteAddress = {
    latitude: number,
    longitude: number,
    geometry: [Object],
    country: string,
    countryCode: string,
    countryFlag: string,
    county: string,
    distance: number,
    city: string,
    number: string,
    postalCode: string,
    stateCode: string,
    state: string,
    street: string,
    layer: string,
    formattedAddress: string,
    addressLabel: string
}
  